Unify command_interactive into a parameter inside the command tool
